rename: console 실행과 web 실행 패키지 구분
